requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
openai==1.75.0
aiohttp==3.13.1
pydantic==2.12.3
python-dotenv==1.1.0